+++ src/main/java/com/example/demo/stock/dto/res/RoundResultResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Builder
 public record RoundResultResponse(
         Integer roundNumber,
-        Long year,
+        long year,
         List<TeamInvestmentDto> teamInvestments
 ) {
 
@@ -15,7 +15,7 @@ public record RoundResultResponse(
     public record TeamInvestmentDto(
             String teamName,
             String maxInvestmentStock,
-            Integer totalInvestmentAmount
+            long totalInvestmentAmount
     ) {
     }
 }
